pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trafikwatch"
version = "0.3.0"
description = "Real-time SNMP interface monitoring from the terminal"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "pysnmp>=7.1,<8",
    "pyyaml>=6.0,<7",
    "rich>=14.0,<15",
    "textual>=7.0,<8",
    "textual-plotext>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-cov>=4.1",
]

[project.scripts]
trafikwatch = "trafikwatch.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["trafikwatch"]